Fix socket close exception being mixed up with a socket dispose.
Previously, closing the socket would cause the SocketException in ReadInternal to be dropped, bubbling up in finalizer. This is because Socket.EndRead() would throw ObjectDisposedException and drop the real exception instead, when trying to handle the error. The HttpListener code still *worked* (since it had a blanket try-catch-all) but the finalizer exception was bad. To fix this I just moved the read code from APM to TAP so the exception could be retrieved properly.