Removes 411 status code for POST requests with an empty body.
Removes 411 status code for POST requests with an empty body.

This seems dumb, but some systems does that, and it's a mistake to not let them do their weird requests.

An example would be the SVO (Sports View Online) server used in some PlayStation games which sends a POST request with nothing inside.