Implements back the port based SSL certificate fetching.
Implements back the port based SSL certificate fetching.

This modernize the Mono code base to allow booth, a in-memory cache and that same file-based cache if the in-memory cache doesn't contain the entry.

This makes the previous SetCertificate constructor obsolete, but for good, since we can now set a different certificate per ports.